# docs/courier/fxcouriers/v1.0.0/external-api-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KurierSystem REST API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://fxcouriers.kuriersystem.pl/api/rest"
    }
  ],
  "paths": {
...
    "/services": {
      "get": {
        "tags": ["basic"],
        "responses": { "200": { "description": "OK" } },
        "security": [{ "bearer_auth": [] }]
      }
    },
    "/company/{company_id}": {
      "get": {
        "tags": ["company"],
        "parameters": [
          { "name": "company_id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int32" } }
        ],
        "responses": { "200": { "description": "OK" } },
        "security": [{ "bearer_auth": [] }]
      },
      "put": {
        "tags": ["company"],
        "parameters": [
          { "name": "company_id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int32" } }
        ],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Company" } } }
        },
        "responses": { "200": { "description": "OK" } },
        "security": [{ "bearer_auth": [] }]
      }
    },
...
    "/shipments": {
      "post": {
        "tags": ["shipment"],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Shipment" } } }
        },
        "responses": { "200": { "description": "OK" } },
        "security": [{ "bearer_auth": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "country": { "type": "string" },
          "city": { "type": "string" },
          "postal_code": { "type": "string" },
          "street": { "type": "string" },
          "house_number": { "type": "integer", "format": "int32" },
          "apartment_number": { "type": "integer", "format": "int32" },
          "contact_person": { "type": "string" },
          "contact_phone": { "type": "string" },
          "contact_email": { "type": "string" },
          "services": { "type": "array", "items": { "$ref": "#/components/schemas/Service" } }
        }
      },
      "Company": {
        "type": "object",
        "properties": {
          "company_id": { "type": "integer", "format": "int32" },
          "name": { "type": "string" },
          "short_name": { "type": "string" },
          "description": { "type": "string" },
          "taxno": { "type": "string" },
          "taxno_additional": { "type": "string" },
          "currency": { "type": "string" },
          "www": { "type": "string" },
          "is_active": { "type": "boolean" },
          "address": { "$ref": "#/components/schemas/Address" }
        }
      },
...
      "Service": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "value": { "type": "string" },
          "quantity": { "type": "integer", "format": "int32" }
        }
      },
      "Shipment": {
        "type": "object",
        "properties": {
          "pickup_date": { "type": "string" },
          "pickup_time_from": { "type": "string" },
          "pickup_time_to": { "type": "string" },
          "order_id_list": { "type": "array", "format": "int32", "items": { "type": "integer" } }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  }
}